requirements.txt

streamlit>=1.50.0
openai>=1.0.0
requests>=2.27.0
python-docx>=0.8.11
python-dotenv>=1.0.0
fastapi
uvicorn
